+++ lib/mtbuild/application_configuration.rb
@@ -18,10 +18,10 @@ module MTBuild
         all_object_folders |= object_folders
       end
 
-      application_binaries, application_files, application_folders = @default_toolchain.create_application_tasks(all_object_files, @project_name)
+      application_binaries, application_files, application_folders = @default_toolchain.create_application_tasks(all_object_files, @parent_project.project_name)
       dependencies = @dependencies+all_object_folders+application_folders+application_files+application_binaries
 
-      desc "Build application '#{@project_name}' with configuration '#{@configuration_name}'"
+      desc "Build application '#{@parent_project.project_name}' with configuration '#{@configuration_name}'"
       new_task = application_task @configuration_name => dependencies do |t|
         puts "built application #{t.name}."
         @tests.each do |test|
@@ -34,12 +34,12 @@ module MTBuild
       end
 
       namespace @configuration_name do
-        OrganizedPackageTask.new("#{@project_name}-#{@configuration_name}", :noversion) do |t|
+        OrganizedPackageTask.new("#{@parent_project.project_name}-#{@configuration_name}", :noversion) do |t|
           t.need_tar_gz = true
           t.add_files_to_folder("", application_binaries+application_files)
         end
       end
-      Rake::Task[:"#{@project_name}:#{@configuration_name}:package"].prerequisites.insert(0, :"#{@project_name}:#{@configuration_name}")
+      Rake::Task[:"#{@parent_project.project_name}:#{@configuration_name}:package"].prerequisites.insert(0, :"#{@parent_project.project_name}:#{@configuration_name}")
 
     end