+++ lib/mtbuild/configuration.rb
@@ -8,8 +8,8 @@ module MTBuild
     # The configuration's name
     attr_reader :configuration_name
 
-    # The name of the project that owns this configuration
-    attr_reader :project_name
+    # The project that owns this configuration
+    attr_reader :parent_project
 
     # The project's folder. Relative path references are interpreted as
     # relative to this folder.
@@ -18,16 +18,16 @@ module MTBuild
     # The project's output folder. Project output goes here.
     attr_reader :output_folder
 
-    def initialize(project_name, project_folder, output_folder, configuration_name, configuration)
-      check_configuration(configuration)
-      @project_name = project_name
-      @project_folder = File.expand_path(project_folder)
+    def initialize(parent_project, output_folder, configuration_name, configuration)
+      @parent_project = parent_project
+      @project_folder = File.expand_path(@parent_project.project_folder)
       @configuration_name = configuration_name
-      @output_folder = File.expand_path(File.join(output_folder, @project_name.to_s, @configuration_name.to_s))
+      @output_folder = File.expand_path(File.join(output_folder, @configuration_name.to_s))
+      check_configuration(configuration)
 
       @versioner = nil
       @versioner_config = configuration.fetch(:versioner, nil)
-      @versioner = Versioner.create_versioner(@project_name, @project_folder, @output_folder, @configuration_name, @versioner_config) unless @versioner_config.nil?
+      @versioner = Versioner.create_versioner(@parent_project.project_name, @project_folder, @output_folder, @configuration_name, @versioner_config) unless @versioner_config.nil?
     end
 
     # Create the actual Rake tasks that will perform the configuration's work
@@ -38,6 +38,11 @@ module MTBuild
     private
 
     def check_configuration(configuration)
+    end
+
+    def namespace_tasks(task_list)
+      return task_list.collect {|task| "#{@parent_project.parent_workspace.workspace_name}:#{task}"} unless @parent_project.parent_workspace.nil?
+      return task_list
     end
 
     include Rake::DSL