+++ lib/mtbuild/framework_project.rb
@@ -16,9 +16,10 @@ module MTBuild
     # Adds a named FrameworkConfiguration to the project.
     def add_configuration(configuration_name, configuration)
       super
-      default_configuration = Workspace.configuration_defaults.fetch(configuration_name, {})
+      default_configuration = {}
+      default_configuration = @parent_workspace.configuration_defaults.fetch(configuration_name, {}) unless @parent_workspace.nil?
       merged_configuration = Utils.merge_configurations(default_configuration, configuration)
-      cfg = FrameworkConfiguration.new(@project_name, @project_folder, effective_output_folder, configuration_name, merged_configuration, @api_headers)
+      cfg = FrameworkConfiguration.new(self, effective_output_folder, configuration_name, merged_configuration, @api_headers)
       @configurations << cfg
       return cfg
     end