+++ lib/mtbuild/project.rb
@@ -2,6 +2,7 @@ module MTBuild
 
   # This is the base class for all project types.
   class Project
+    require 'mtbuild/build_registry'
 
     # The project's name
     attr_reader :project_name
@@ -13,29 +14,38 @@ module MTBuild
     # The project's output folder. Project output goes here.
     attr_reader :output_folder
 
+    # The project's parent workspace
+    attr_reader :parent_workspace
+
     # If supplied, the configuration_block will be passed the
     # newly-constructed Project object.
     def initialize(project_name, project_folder, &configuration_block)
       @configurations = []
-      @default_tasks = []
-      @project_name = project_name
       @project_folder = File.expand_path(project_folder)
       @output_folder = File.expand_path(File.join(@project_folder, MTBuild.default_output_folder))
+      @project_name, @parent_workspace = MTBuild::BuildRegistry.enter_project(project_name, self)
+
       configuration_block.call(self) if configuration_block
 
       namespace @project_name do
         @configurations.each do |configuration|
-          configuration.configure_tasks()
+          configuration.configure_tasks
         end
       end
 
-      # If there is no active workspace, set up any registered default project tasks
-      task :default => @default_tasks unless Workspace.have_workspace?
+      #TODO: This is a strange way to do this. Should probably be moved to "application" functionality.
+      if @parent_workspace.nil? and Rake.application.lookup(:default).nil?
+        task :default do
+          puts 'Nothing to do. Use the -T flag to see the list of tasks you can build.'
+        end
+      end
+
+      MTBuild::BuildRegistry.exit_project
     end
 
-    # Add tasks to be built by default if MTBuild is invoked with no arguments
-    def add_default_tasks(default_tasks)
-      @default_tasks |= Utils.ensure_array(default_tasks).flatten
+    # Get the fully-qualified task name for a configuration
+    def task_for_configuration(config_name)
+      "#{@project_name}:#{config_name}"
     end
 
     # Set the project's output folder.
@@ -47,12 +57,10 @@ module MTBuild
     # return the workspace's output folder. If not, it will return the
     # project's output folder.
     def effective_output_folder
-      if Workspace.have_workspace?
-        relative_project_location = Utils::path_difference(Workspace.workspace.workspace_folder, @project_folder)
-        fail "Project folder '#{@project_folder}' must be within workspace folder '#{Workspace.workspace.workspace_folder}'." if relative_project_location.nil?
-        return File.join(Workspace.workspace.output_folder, relative_project_location)
+      if MTBuild::BuildRegistry.top_workspace.nil?
+        File.join(@output_folder, @project_name.to_s.split(':'))
       else
-        return @output_folder
+        File.join(MTBuild::BuildRegistry.top_workspace.output_folder, @project_name.to_s.split(':'))
       end
     end