+++ lib/mtbuild/workspace.rb
@@ -3,6 +3,7 @@ module MTBuild
   # Use this class to create a workspace
   class Workspace
     require 'mtbuild/utils'
+    require 'mtbuild/build_registry'
     require 'rake/clean'
 
     # The workspace's name
@@ -15,32 +16,60 @@ module MTBuild
     # The workspace's output folder
     attr_reader :output_folder
 
+    # The workspace's parent workspace
+    attr_reader :parent_workspace
+
+    # Workspace configuration defaults
+    attr_reader :configuration_defaults
+
     def initialize(workspace_name, workspace_folder, &configuration_block)
-      @workspace_name = workspace_name
       @workspace_folder = File.expand_path(workspace_folder)
       @output_folder = File.expand_path(File.join(@workspace_folder, MTBuild.default_output_folder))
       @projects = []
+      @workspaces = []
       @default_tasks = []
+      @configuration_defaults = {}
+      @child_workspaces = {}
+      @workspace_name, @parent_workspace = MTBuild::BuildRegistry.enter_workspace(workspace_name, self)
 
-      # Only process the first workspace found.
-      # Subsequent workspaces will be ignored
-      if MTBuild::Workspace.workspace.nil?
-        MTBuild::Workspace.set_workspace(self)
+      configuration_block.call(self) if configuration_block
 
-        configuration_block.call(self) if configuration_block
+      # If there's a parent workspace, use its output folder.
+      # Don't use the current workspace's output folder.
+      @output_folder = @parent_workspace.output_folder unless @parent_workspace.nil?
 
-        @projects.each do |project|
-          require project
-        end
-
-        CLOBBER.include(@output_folder)
-
-        task :workspace do
-          puts "built workspace #{@workspace_name}"
-        end
-
-        task :default => @default_tasks+[:workspace]
+      @workspaces.each do |workspace|
+        MTBuild::BuildRegistry.expect_workspace
+        require workspace
       end
+
+      @projects.each do |project|
+        MTBuild::BuildRegistry.expect_project
+        require project
+      end
+
+      CLOBBER.include(@output_folder)
+
+      task :workspace do
+        puts "built workspace #{@workspace_name}"
+      end
+
+      task :default => @default_tasks+[:workspace]
+
+      MTBuild::BuildRegistry.exit_workspace
+    end
+
+    # Adds a workspace subfolder
+    def add_workspace(workspace_location)
+      new_workspaces = []
+      Utils.expand_folder_list(workspace_location, Rake.original_dir).each do |workspace_path|
+        if File.directory? workspace_path
+          workspace_rakefile = MTBuild::Workspace.find_build_file(workspace_path)
+          new_workspaces << workspace_rakefile unless workspace_rakefile.nil?
+        end
+      end
+      $stderr.puts "Could not find a valid workspace at '#{workspace_location}'. Ignored." if new_workspaces.empty?
+      @workspaces += new_workspaces
     end
 
     # Adds a project subfolder
@@ -48,7 +77,7 @@ module MTBuild
       new_projects = []
       Utils.expand_folder_list(project_location, Rake.original_dir).each do |project_path|
         if File.directory? project_path
-          project_rakefile = MTBuild::Workspace.find_mtbuildfile(project_path)
+          project_rakefile = MTBuild::Workspace.find_build_file(project_path)
           new_projects << project_rakefile unless project_rakefile.nil?
         end
       end
@@ -56,14 +85,22 @@ module MTBuild
       @projects += new_projects
     end
 
-    # Adds tasks to be run by default when MTBuild is invoked with no arguments.
+    # Adds tasks to be run by default when MTBuild is invoked with no arguments. This method
+    # expects only MTBuild tasks and will namespace them according to the current workspace
+    # hierarchy.
     def add_default_tasks(default_tasks)
+      @default_tasks |= Utils.ensure_array(default_tasks).flatten.collect {|default_task| "#{@workspace_name}:#{default_task}"}
+    end
+
+    # Adds Rake tasks to be run by default when MTBuild is invoked with no arguments. This method
+    # will not namespace the tasks, so it can be used to specify plain Rake task names.
+    def add_default_rake_tasks(default_tasks)
       @default_tasks |= Utils.ensure_array(default_tasks).flatten
     end
 
     # Sets defaults for all configurations with the specified name
     def set_configuration_defaults(configuration_name, defaults_hash)
-      Workspace.set_configuration_defaults(configuration_name, defaults_hash)
+      @configuration_defaults[configuration_name] = defaults_hash
     end
 
     # Sets the build output folder location
@@ -71,45 +108,16 @@ module MTBuild
       @output_folder = File.expand_path(File.join(@workspace_folder,output_folder))
     end
 
-    @workspace = nil
-
-    # The singleton workspace instance
-    def self.workspace
-      return @workspace
-    end
-
-    # Sets the singleton workspace instance
-    def self.set_workspace(workspace)
-      @workspace = workspace
-    end
-
-    # Queries whether a workspace exists
-    def self.have_workspace?
-      return !@workspace.nil?
-    end
-
-    # Add default tasks to the singleton workspace instance
+    # Add default tasks to the last active workspace
     def self.add_default_tasks(default_tasks)
-      @workspace.add_default_tasks(default_tasks) unless @workspace.nil?
+      MTBuild::BuildRegistry.last_active_workspace.add_default_rake_tasks(default_tasks) unless MTBuild::BuildRegistry.last_active_workspace.nil?
     end
 
-    @configuration_defaults = {}
-
-    # Gets the singleton configuration defaults
-    def self.configuration_defaults
-      return @configuration_defaults
-    end
-
-    # Adds or updates defaults to the singleton configuration defaults
-    def self.set_configuration_defaults(configuration_name, defaults_hash)
-      @configuration_defaults[configuration_name] = defaults_hash
-    end
-
-    def self.find_mtbuildfile(project_path)
+    def self.find_build_file(project_path)
       Rake.application.rakefiles.each do |fn|
-        mtbuildfile = File.join(project_path, fn)
-        if File.file? mtbuildfile
-          return mtbuildfile
+        build_file = File.join(project_path, fn)
+        if File.file? build_file
+          return build_file
         end
       end
       return nil